requirements.txt
selenium>=4.23
webdriver-manager>=4.0.2
pytest>=8.2
pytest-html>=4.1.1
pytest-xdist>=3.6
python-dotenv>=1.0.1
PyYAML>=6.0.2
